ブログを削除する処理にユーザーの判定機能を追加した

+++ app/Http/Controllers/BlogController.php
@@ -58,6 +58,12 @@ class BlogController extends Controller
      */
     public function destroy($id)
     {
+        $login_user_id = Auth::id();
+        $blog_user_id = Blog::find($id)->user_id;
+        if ($login_user_id !== $blog_user_id) {
+            abort(403);
+        }
+
         $blog = Blog::find($id);
         $blog->delete();
         return redirect()->route('blog.create');
